implement message queue for background tasks
When a long action takes place, we spawn a new thread to do the work so
that it doesn't block the UI. When finished, it pushes a Message into a
thread-safe queue which is read by the main thread at the start of every
frame.

Also introduced a block_task field that drops all event handling when
it's not null.